// mobile-preview-archive/packages/client/account/src/types.ts
export interface TokenPair {
  readonly access_token: string
  readonly refresh_token: string | null
  readonly token_type: string
  readonly expires_in: number
}